refactor: move cells to upper level folder
