+++ frontend/appflowy_flutter/lib/plugins/database_view/widgets/row/cells/cell_container.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:styled_widget/styled_widget.dart';
 
-import '../../layout/sizes.dart';
-import '../row/row.dart';
-import 'cell_accessory.dart';
+import '../../../grid/presentation/layout/sizes.dart';
+import '../../../grid/presentation/widgets/row/row.dart';
+import '../accessory/cell_accessory.dart';
+import '../accessory/cell_shortcuts.dart';
 import 'cell_builder.dart';
-import 'cell_shortcuts.dart';
 
 class CellContainer extends StatelessWidget {
   final GridCellWidget child;
