+++ frontend/appflowy_flutter/lib/plugins/database_view/widgets/row/cells/number_cell.dart
@@ -4,8 +4,8 @@ import 'package:appflowy/startup/startup.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../application/cell/number_cell_bloc.dart';
-import '../../layout/sizes.dart';
+import '../../../grid/application/cell/number_cell_bloc.dart';
+import '../../../grid/presentation/layout/sizes.dart';
 import 'cell_builder.dart';
 
 class GridNumberCell extends GridCellWidget {
